Just rename `runtime_common_*` files to correct `polkadot_runtime_common_*`
