+++ relay/polkadot/src/weights/mod.rs
@@ -42,12 +42,12 @@ pub mod pallet_utility;
 pub mod pallet_vesting;
 pub mod pallet_whitelist;
 pub mod pallet_xcm;
+pub mod polkadot_runtime_common_auctions;
+pub mod polkadot_runtime_common_claims;
+pub mod polkadot_runtime_common_crowdloan;
 pub mod polkadot_runtime_common_identity_migrator;
-pub mod runtime_common_auctions;
-pub mod runtime_common_claims;
-pub mod runtime_common_crowdloan;
-pub mod runtime_common_paras_registrar;
-pub mod runtime_common_slots;
+pub mod polkadot_runtime_common_paras_registrar;
+pub mod polkadot_runtime_common_slots;
 pub mod runtime_parachains_configuration;
 pub mod runtime_parachains_disputes;
 pub mod runtime_parachains_disputes_slashing;
